Pull shared members up to base class
